pyproject.toml
[tool.poetry]
name = "ocr-bot"
version = "0.1.0"
description = ""
authors = ["urm8 <megafukz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
rarfile = "^3.1"
pillow = "^7.1.2"
pytesseract = "^0.3.4"
python-docx = "^0.8.10"
python-telegram-bot = {extras = ["socks"], version = "^12.6.1"}
pysocks = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
